Refactor user-related models into a separate file.
