+++ internals/models_test.py
@@ -538,102 +538,6 @@ class CommentTest(testing_config.CustomTestCase):
     self.assertEqual([], actual_3)
 
 
-class UserPrefTest(testing_config.CustomTestCase):
-
-  def setUp(self):
-    self.user_pref_1 = models.UserPref(email='one@example.com')
-    self.user_pref_1.notify_as_starrer = False
-    self.user_pref_1.put()
-
-    self.user_pref_2 = models.UserPref(email='two@example.com')
-    self.user_pref_2.put()
-
-  def tearDown(self):
-    self.user_pref_1.key.delete()
-    self.user_pref_2.key.delete()
-
-  # @mock.patch('google.appengine.api.users.get_current_user')
-  @mock.patch('framework.users.get_current_user')
-  def test_get_signed_in_user_pref__anon(self, mock_get_current_user):
-    mock_get_current_user.return_value = None
-    actual = models.UserPref.get_signed_in_user_pref()
-    self.assertIsNone(actual)
-
-  # @mock.patch('google.appengine.api.users.get_current_user')
-  @mock.patch('framework.users.get_current_user')
-  def test_get_signed_in_user_pref__first_time(self, mock_get_current_user):
-    mock_get_current_user.return_value = testing_config.Blank(
-        email=lambda: 'user1@example.com')
-
-    actual = models.UserPref.get_signed_in_user_pref()
-
-    self.assertEqual('user1@example.com', actual.email)
-    self.assertEqual(True, actual.notify_as_starrer)
-    self.assertEqual(False, actual.bounced)
-
-  # @mock.patch('google.appengine.api.users.get_current_user')
-  @mock.patch('framework.users.get_current_user')
-  def test_get_signed_in_user_pref__had_pref(self, mock_get_current_user):
-    mock_get_current_user.return_value = testing_config.Blank(
-        email=lambda: 'user2@example.com')
-    user_pref = models.UserPref(
-        email='user2@example.com', notify_as_starrer=False, bounced=True)
-    user_pref.put()
-
-    actual = models.UserPref.get_signed_in_user_pref()
-
-    self.assertEqual('user2@example.com', actual.email)
-    self.assertEqual(False, actual.notify_as_starrer)
-    self.assertEqual(True, actual.bounced)
-
-  # @mock.patch('google.appengine.api.users.get_current_user')
-  @mock.patch('framework.users.get_current_user')
-  def test_dismiss_cue(self, mock_get_current_user):
-    """We store the fact that a user has dismissed a cue card."""
-    mock_get_current_user.return_value = testing_config.Blank(
-        email=lambda: 'one@example.com')
-
-    models.UserPref.dismiss_cue('welcome-message')
-
-    revised_user_pref = models.UserPref.get_signed_in_user_pref()
-    self.assertEqual('one@example.com', revised_user_pref.email)
-    self.assertEqual(['welcome-message'], revised_user_pref.dismissed_cues)
-
-  # @mock.patch('google.appengine.api.users.get_current_user')
-  @mock.patch('framework.users.get_current_user')
-  def test_dismiss_cue__double(self, mock_get_current_user):
-    """We ignore the same user dismissing the same cue multiple times."""
-    mock_get_current_user.return_value = testing_config.Blank(
-        email=lambda: 'one@example.com')
-
-    models.UserPref.dismiss_cue('welcome-message')
-    models.UserPref.dismiss_cue('welcome-message')
-
-    revised_user_pref = models.UserPref.get_signed_in_user_pref()
-    self.assertEqual('one@example.com', revised_user_pref.email)
-    self.assertEqual(['welcome-message'], revised_user_pref.dismissed_cues)
-
-  def test_get_prefs_for_emails__some_found(self):
-    emails = ['one@example.com', 'two@example.com', 'huh@example.com']
-    user_prefs = models.UserPref.get_prefs_for_emails(emails)
-    self.assertEqual(3, len(user_prefs))
-    one, two, huh = user_prefs
-    self.assertEqual('one@example.com', one.email)
-    self.assertFalse(one.notify_as_starrer)
-    self.assertEqual('two@example.com', two.email)
-    self.assertTrue(two.notify_as_starrer)
-    # This one is automatically created:
-    self.assertEqual('huh@example.com', huh.email)
-    self.assertTrue(huh.notify_as_starrer)
-
-  def test_get_prefs_for_emails__long_list(self):
-    emails = ['user_%d@example.com' % i
-              for i in range(100)]
-    user_prefs = models.UserPref.get_prefs_for_emails(emails)
-    self.assertEqual(100, len(user_prefs))
-    self.assertEqual('user_0@example.com', user_prefs[0].email)
-
-
 class OwnersFileTest(testing_config.CustomTestCase):
 
   def setUp(self):
