+++ pages/guideforms.py
@@ -24,12 +24,11 @@ from django.core.exceptions import ValidationError
 from django.utils.html import conditional_escape, escape
 from django.utils.safestring import mark_safe
 
-# from google.appengine.api import users
 from framework import users
-
 from internals import core_enums
-from internals import models
 from internals import processes
+from internals import user_models
+import settings
 
 
 # This is the longest string that a cloud ndb StringProperty seems to accept.
@@ -448,8 +447,9 @@ ALL_FIELDS = {
 
     'blink_components': forms.ChoiceField(
       required=False, label='',
-      choices=[(x, x) for x in models.BlinkComponent.fetch_all_components()],
-      initial=[models.BlinkComponent.DEFAULT_COMPONENT]),
+      choices=[(x, x) for x in
+               user_models.BlinkComponent.fetch_all_components()],
+      initial=[settings.DEFAULT_COMPONENT]),
 
     'devrel': MultiEmailField(
         required=False, label='',
